ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
opencv-python==4.9.0.80
insightface==0.7.3
ultralytics==8.1.0
numpy==1.26.4
Pillow==10.3.0
psycopg2-binary==2.9.9
pgvector==0.2.5
pymongo==4.6.3
cloudinary==1.39.1
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0
onnxruntime==1.17.3